Refactor probe-core(-socks) for robustness, 1/n
This code is untested and the Python-side changes required to make it work
have not yet been written.  The essential idea is that instead of reading and
writing textual results, both programs now take a fd pointing at a shared
memory segment; they receive all their input from there and they deposit
their output there as well.  Thus, if the C program crashes in the middle,
work done up to that point is still visible to the Python script, and can
be continued in a new instance from where it left off.

Most of the duplicated code between the two variants of the program is now
also factored out to a shared .c file.  The main loop is still rather
redundant, though.