Replace PHP server scripts with Flask-based Python.
This reduces the number of programming languages involved in this
project and will make it easier to do fancier things on the server
side.  One longstanding bug is already fixed in the process: we can
now do lookups in the MaxMind database ourselves instead of invoking a
third-party web service that may be blocked by anti-ad/anti-tracking
browser policies.